refactor: extract beads-rust client to match beads pattern
Move br CLI wrapping logic from the tracker adapter into a dedicated
src/beads-rust/client.ts, mirroring the src/beads/client.ts pattern.
The tracker adapter is now a thin layer that consumes the client.
